test(rpc): extract metrics based api versioning test
