+++ crates/rpc/src/context.rs
@@ -37,12 +37,12 @@ impl RpcContext {
         }
     }
 
-    #[cfg(test)]
+    #[cfg(any(test, feature = "test-utils"))]
     pub fn for_tests() -> Self {
         Self::for_tests_on(pathfinder_common::Chain::Testnet)
     }
 
-    #[cfg(test)]
+    #[cfg(any(test, feature = "test-utils"))]
     pub fn for_tests_on(chain: pathfinder_common::Chain) -> Self {
         assert_ne!(chain, Chain::Mainnet, "Testing on MainNet?");
 
@@ -55,13 +55,13 @@ impl RpcContext {
             Chain::Custom => unreachable!("Should not be testing with custom chain"),
         };
 
-        let storage = super::tests::setup_storage();
+        let storage = super::test_utils::setup_storage();
         let sync_state = Arc::new(SyncState::default());
         let sequencer = SequencerClient::new(chain).unwrap();
         Self::new(storage, sync_state, chain_id, sequencer)
     }
 
-    #[cfg(test)]
+    #[cfg(any(test, feature = "test-utils"))]
     pub fn with_storage(self, storage: Storage) -> Self {
         Self { storage, ..self }
     }
@@ -73,12 +73,12 @@ impl RpcContext {
         }
     }
 
-    #[cfg(test)]
+    #[cfg(any(test, feature = "test-utils"))]
     pub async fn for_tests_with_pending() -> Self {
         // This is a bit silly with the arc in and out, but since its for tests the ergonomics of
         // having Arc also constructed is nice.
         let context = Self::for_tests();
-        let pending_data = super::tests::create_pending_data(context.storage.clone()).await;
+        let pending_data = super::test_utils::create_pending_data(context.storage.clone()).await;
         context.with_pending_data(pending_data)
     }
