+++ crates/rpc/src/versioning.rs
@@ -9,7 +9,7 @@ use jsonrpsee::types::Id;
 use tower::BoxError;
 
 #[derive(thiserror::Error, Debug)]
-pub enum VersioningError {
+enum VersioningError {
     #[error("Invalid path")]
     InvalidPath,
     #[error("Too large: {0}")]
@@ -31,7 +31,7 @@ impl VersioningError {
     }
 }
 
-pub async fn prefix_rpc_method_names_with_version(
+pub(crate) async fn prefix_rpc_method_names_with_version(
     request: Request<Body>,
     max_request_body_size: u32,
 ) -> Result<Request<Body>, BoxError> {
@@ -101,7 +101,7 @@ pub async fn prefix_rpc_method_names_with_version(
     Ok(request)
 }
 
-pub fn try_map_errors_to_responses(
+pub(crate) fn try_map_errors_to_responses(
     result: Result<Response<Body>, BoxError>,
 ) -> Result<Response<Body>, BoxError> {
     match result {
@@ -175,133 +175,50 @@ mod response {
     }
 }
 
+pub mod test_utils {
+    pub mod method_names {
+        pub const COMMON_FOR_V02_V03: [&str; 23] = [
+            "starknet_addDeclareTransaction",
+            "starknet_addDeployAccountTransaction",
+            "starknet_addInvokeTransaction",
+            "starknet_blockHashAndNumber",
+            "starknet_blockNumber",
+            "starknet_call",
+            "starknet_chainId",
+            "starknet_estimateFee",
+            "starknet_getBlockWithTxHashes",
+            "starknet_getBlockWithTxs",
+            "starknet_getBlockTransactionCount",
+            "starknet_getClass",
+            "starknet_getClassAt",
+            "starknet_getClassHashAt",
+            "starknet_getEvents",
+            "starknet_getNonce",
+            "starknet_getStateUpdate",
+            "starknet_getStorageAt",
+            "starknet_getTransactionByBlockIdAndIndex",
+            "starknet_getTransactionByHash",
+            "starknet_getTransactionReceipt",
+            "starknet_pendingTransactions",
+            "starknet_syncing",
+        ];
+        pub const COMMON_FOR_ALL: [&str; 2] =
+            ["pathfinder_getProof", "pathfinder_getTransactionStatus"];
+        pub const V03_ONLY: [&str; 1] = ["starknet_simulateTransaction"];
+        pub const PATHFINDER_ONLY: [&str; 1] = ["pathfinder_version"];
+    }
+
+    pub mod paths {
+        pub const V02: &[&str] = &["", "/", "/rpc/v0.2", "/rpc/v0.2/"];
+        pub const V03: &[&str] = &["/rpc/v0.3", "/rpc/v0.3/"];
+        pub const PATHFINDER: &[&str] = &["/rpc/pathfinder/v0.1", "/rpc/pathfinder/v0.1/"];
+    }
+}
+
 #[cfg(test)]
 mod tests {
 
-    const COMMON_FOR_V02_V03: [&str; 23] = [
-        "starknet_addDeclareTransaction",
-        "starknet_addDeployAccountTransaction",
-        "starknet_addInvokeTransaction",
-        "starknet_blockHashAndNumber",
-        "starknet_blockNumber",
-        "starknet_call",
-        "starknet_chainId",
-        "starknet_estimateFee",
-        "starknet_getBlockWithTxHashes",
-        "starknet_getBlockWithTxs",
-        "starknet_getBlockTransactionCount",
-        "starknet_getClass",
-        "starknet_getClassAt",
-        "starknet_getClassHashAt",
-        "starknet_getEvents",
-        "starknet_getNonce",
-        "starknet_getStateUpdate",
-        "starknet_getStorageAt",
-        "starknet_getTransactionByBlockIdAndIndex",
-        "starknet_getTransactionByHash",
-        "starknet_getTransactionReceipt",
-        "starknet_pendingTransactions",
-        "starknet_syncing",
-    ];
-    const COMMON_FOR_ALL: [&str; 2] = ["pathfinder_getProof", "pathfinder_getTransactionStatus"];
-    const V03_ONLY: [&str; 1] = ["starknet_simulateTransaction"];
-    const PATHFINDER_ONLY: [&str; 1] = ["pathfinder_version"];
-
-    const V02_PATHS: &[&str] = &["", "/", "/rpc/v0.2", "/rpc/v0.2/"];
-    const V03_PATHS: &[&str] = &["/rpc/v0.3", "/rpc/v0.3/"];
-    const PATHFINDER_PATHS: &[&str] = &["/rpc/pathfinder/v0.1", "/rpc/pathfinder/v0.1/"];
-
-    #[tokio::test]
-    async fn api_versions_are_routed_correctly_for_all_methods() {
-        use crate::test_client::TestClientBuilder;
-        use crate::{RpcContext, RpcMetricsLogger, RpcServer};
-        use pathfinder_common::test_utils::metrics::{FakeRecorder, RecorderGuard};
-        use serde_json::json;
-
-        let context = RpcContext::for_tests();
-        let (_server_handle, address) = RpcServer::new("127.0.0.1:0".parse().unwrap(), context)
-            .with_logger(RpcMetricsLogger)
-            .run()
-            .await
-            .unwrap();
-
-        let v02_methods = COMMON_FOR_V02_V03
-            .into_iter()
-            .chain(COMMON_FOR_ALL.into_iter())
-            .collect::<Vec<_>>();
-        let v03_methods = v02_methods
-            .clone()
-            .into_iter()
-            .chain(V03_ONLY.into_iter())
-            .collect::<Vec<_>>();
-        let pathfinder_methods = COMMON_FOR_ALL
-            .into_iter()
-            .chain(PATHFINDER_ONLY.into_iter())
-            .collect();
-
-        for (paths, version, methods) in vec![
-            (V02_PATHS, "v0.2", v02_methods),
-            (V03_PATHS, "v0.3", v03_methods),
-            (PATHFINDER_PATHS, "v0.1", pathfinder_methods),
-        ]
-        .into_iter()
-        {
-            let recorder = FakeRecorder::default();
-            let handle = recorder.handle();
-            // Other concurrent tests could be setting their own recorders
-            let guard = RecorderGuard::lock(recorder);
-
-            let paths_len = paths.len();
-            let paths_iter = paths.iter();
-
-            // Perform all the calls but don't assert the results just yet
-            for path in paths_iter.clone().map(ToOwned::to_owned) {
-                let client = TestClientBuilder::default()
-                    .address(address)
-                    .endpoint(path.into())
-                    .build()
-                    .unwrap();
-
-                for method in methods.iter() {
-                    let res = client.request::<serde_json::Value>(method, json!([])).await;
-
-                    match res {
-                        Err(jsonrpsee::core::Error::Call(
-                            jsonrpsee::types::error::CallError::Custom(e),
-                        )) if e.code() == jsonrpsee::types::error::METHOD_NOT_FOUND_CODE => {
-                            // Don't poison the internal lock
-                            drop(guard);
-                            panic!("Unregistered method called, path: {path}, method: {method}")
-                        }
-                        Ok(_) | Err(_) => {}
-                    }
-                }
-            }
-
-            // Drop the global recorder guard to avoid poisoning its internal lock if
-            // the following asserts fail which would fail other tests using the `RecorderGuard`
-            // at the same time.
-            //
-            // The recorder itself still exists since dropping the guard only unregisters the recorder
-            // and leaks it making the handle still valid past this point.
-            drop(guard);
-
-            // Now we can safely assert all results
-            for path in paths_iter.clone() {
-                for method in methods.iter() {
-                    let expected_counter = paths_len as u64;
-                    let actual_counter = handle.get_counter_value_by_label(
-                        "rpc_method_calls_total",
-                        [("method", method), ("version", version)],
-                    );
-                    assert_eq!(
-                        actual_counter, expected_counter,
-                        "path: {path}, method: {method}"
-                    );
-                }
-            }
-        }
-    }
+    use super::test_utils::{method_names, paths};
 
     // In an unintentional way OFC: if a method is INTENDED to be available
     // on many paths then this is absolutely allowed.
@@ -317,21 +234,23 @@ mod tests {
             .await
             .unwrap();
 
-        let not_in_v02 = V03_ONLY
+        let not_in_v02 = method_names::V03_ONLY
             .into_iter()
             .clone()
-            .chain(PATHFINDER_ONLY.into_iter())
+            .chain(method_names::PATHFINDER_ONLY.into_iter())
             .collect::<Vec<_>>();
-        let not_in_v03 = PATHFINDER_ONLY.into_iter().collect::<Vec<_>>();
-        let not_in_pathfinder = COMMON_FOR_V02_V03
+        let not_in_v03 = method_names::PATHFINDER_ONLY
             .into_iter()
-            .chain(V03_ONLY.into_iter())
+            .collect::<Vec<_>>();
+        let not_in_pathfinder = method_names::COMMON_FOR_V02_V03
+            .into_iter()
+            .chain(method_names::V03_ONLY.into_iter())
             .collect::<Vec<_>>();
 
         for (paths, methods) in vec![
-            (V02_PATHS, not_in_v02),
-            (V03_PATHS, not_in_v03),
-            (PATHFINDER_PATHS, not_in_pathfinder),
+            (paths::V02, not_in_v02),
+            (paths::V03, not_in_v03),
+            (paths::PATHFINDER, not_in_pathfinder),
         ]
         .into_iter()
         {
